Improve QObject::moveToThread doc
Change mentions of parameter value being "zero" to nullptr.

Clarify that when nullptr is passed to moveToThread() event processing
is stopped because the object is no longer associated with any thread.

Also, reitarete this fact in the paragraph about processing of new
events.

There's an exception to the rule that QObjects cannot be "pulled" by
moveToThread that is buried in the implementation and not mentioned
in the doc. This information is worth noting explicitly.

Change-Id: I816ff737c48d8057b39e36b566079710aeb8e690
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>